+++ src/main/java/com/bootcamp/Spring/challenge/dto/SellerFollowedProductListDTO.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Getter
 @Setter
 public class SellerFollowedProductListDTO {
-    private long sellerId;
+    private Integer sellerId;
     private String sellerName;
     private List<ProductDTO> posts = new ArrayList<>();
